requirements.txt
tenseal==0.3.14
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.10
langchain-ollama>=0.0.2
langchain-mcp-adapters>=0.0.1
numpy>=1.24.0
streamlit>=1.30.0 